resolved variables conflicts

+++ src/main/java/org/example/delegate/PrepareInvoiceDelegate.java
@@ -7,13 +7,13 @@ import org.camunda.bpm.engine.delegate.JavaDelegate;
 public class PrepareInvoiceDelegate implements JavaDelegate {
     public void execute(DelegateExecution delegateExecution) throws Exception {
         Integer number_of_acceptances = (Integer) delegateExecution.getVariable("number_of_acceptances");
-        Integer payment_info = (Integer) delegateExecution.getVariable("payment_info");
+        Double payment_info = (Double) delegateExecution.getVariable("payment_info");
         String date = (String) delegateExecution.getVariable("date");
         //Manual setting of invoice id
         String invoiceid = (String) delegateExecution.getVariable("openingid");
         invoiceid = invoiceid + date;
 
-        Integer net = number_of_acceptances*payment_info;
+        Double net = number_of_acceptances*payment_info;
         delegateExecution.setVariable("net", net);
         Double tax = net*0.19;
         delegateExecution.setVariable("tax", tax);
